\name{ccplot-methods}
\docType{methods}
\alias{ccplot-methods}
\alias{ccplot}
\alias{ccplot,formula,FLCohort-method}
\title{Catch curves plot method}

\description{
This method provides the catch curves plots based on lattice.
}

\section{Generic Function}{
\describe{
\item{Usage}{ccplot(x, data, \dots)}
\item{Arguments:}{\ }
\describe{
	\item{\ }{\code{x}: \code{"formula"} object}
	\item{\ }{\code{data}: \code{"FLCohort"} object}
	\item{\ }{\code{\dots}: additional, optional arguments if requested}
}
\item{Value}{Catch curves plot}
}}

\section{Methods}{
\describe{
\item{x="formula", data="FLCohort"}{Plots catch curves defined by \code{"formula"}.}
}
}
\seealso{
  \code{\link[FLCore]{FLCohort-class}}, \code{\link{lattice}}
}
\examples{
# not ready yet !
}
\keyword{methods}
